+++ src/main/kotlin/sk/streetofcode/webapi/api/NewsletterService.kt
@@ -1,0 +1,7 @@
+package sk.streetofcode.webapi.api
+
+import sk.streetofcode.webapi.api.request.AddEmailToNewsletterRequest
+
+interface NewsletterService {
+    fun addToNewsletter(userId: String? = null, request: AddEmailToNewsletterRequest)
+}
